Different Pizza Sizes, Scripted Events
Pizzas are now Large, Medium, And Small. The smaller the pizza is, the more points it's worth. Pizza size is an enum in the pizza script (but it's made with 3 prefabs).

We can now create pre defined pizza events via the inspector in the character script. for each event, you can choose how many pizzas you want to spawn, and for each pizza, you can choose the spawn point, size, and direction it's pushed.