Pizza Popping
Each pizza has "lives." Every time a pizza is hit with the correct topping, it loses a life, and when it's life is 0, it gets "popped" into the air. My thought behind the life part is that the larger the pizza is, the more hits it will require to pop.